moved TooltipButton in widgets
