+++ ui/src/views/compute/backup/BackupSchedule.vue
@@ -74,7 +74,7 @@
 <script>
 import { api } from '@/api'
 import { timeZoneName } from '@/utils/timezone'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'BackupSchedule',
