+++ ui/src/views/iam/RuleDelete.vue
@@ -25,7 +25,7 @@
 </template>
 
 <script>
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'RuleDelete',
