+++ ui/src/views/iam/SSLCertificateTab.vue
@@ -68,7 +68,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'SSLCertificate',
