+++ ui/src/views/image/IsoZones.vue
@@ -123,7 +123,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'IsoZones',
