+++ ui/src/views/image/TemplateZones.vue
@@ -132,7 +132,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'TemplateZones',
