+++ ui/src/views/infra/network/DedicatedVLANTab.vue
@@ -134,7 +134,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'DedicatedVLANTab',
