+++ ui/src/views/infra/network/IpRangesTabPublic.vue
@@ -205,7 +205,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'IpRangesTabPublic',
