+++ ui/src/views/infra/network/IpRangesTabStorage.vue
@@ -117,7 +117,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'IpRangesTabStorage',
