+++ ui/src/views/infra/zone/IpAddressRangeForm.vue
@@ -142,7 +142,7 @@
 </template>
 <script>
 
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 import { mixinDevice } from '@/utils/mixin.js'
 
 export default {
