+++ ui/src/views/network/EgressRulesTab.vue
@@ -102,7 +102,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'EgressRulesTab',
