+++ ui/src/views/network/IngressEgressRuleConfigure.vue
@@ -170,7 +170,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   components: {
