+++ ui/src/views/network/InternalLBAssignedVmTab.vue
@@ -63,7 +63,7 @@
 </template>
 <script>
 import { api } from '@/api'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'InternalLBAssignedVmTab',
