+++ ui/src/views/project/AccountsTab.vue
@@ -81,7 +81,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'AccountsTab',
