+++ ui/src/views/project/InvitationsTemplate.vue
@@ -79,7 +79,7 @@
 <script>
 import { api } from '@/api'
 import Status from '@/components/widgets/Status'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 
 export default {
   name: 'InvitationsTemplate',
