+++ ui/src/views/storage/ScheduledSnapshots.vue
@@ -76,7 +76,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipButton from '@/components/view/TooltipButton'
+import TooltipButton from '@/components/widgets/TooltipButton'
 import { timeZoneName } from '@/utils/timezone'
 
 export default {
